Get the site working.
This included deleting a number of things including all of the
Wikipedia pages of the date. I checked and these are in the internet
archive.

This commit also removed the scripts for now. Got the favicon working
and added the frontmatter back (it seems to have been stripped by the
import.