Switch to Gatsby
It feels like most of what I do in this repo is switch static site
generators.

I switched to Hugo pretty recently and was mostly happy with it. A
single executable is nice, but I never really got comfortable with the
templateing system. I have on the other hand been writing React for 4
years now (which isn't a crazy long time, but it mostly just snuck up on
me). I used Gatsby as the static site generator for a friends website
and liked it's development experience a little more.

This commit sets up a basic typescript Gatsby app. It starts the process
of recreating the `hugo-coder` project. I feel a little bit weird about
copying the work but I am not quite sure how to correctly attribute the
work.