Style the footer
The footer was pretty much just my first attempt to throw things
together. This commit updates it to use a flexbox.

I also updated the fonts to use `Inconsolata` which is the font I use in
my terminal (which makes the code blocks in my blog feel like "mine" to
me even though I think it's a common font).