Show which is the currently selected item in the navigation.
I did this by adding a `pathname` prop to the `navigation` and in turn
the `layouts`. This allows the `li` in the `navigation` to use the
`aria-current` attribute. This attribute is documented here:
https://www.w3.org/TR/wai-aria-1.1/#aria-current . The attribute is then
used to add a small underline to indicate that you are already on the
page for the link.