Start to show images for the projects.
I would like the posts to have a little bit more visual appeal then the
blog posts. The thought is to have one per project but right now this is
just the code names project.

The elements in the projects page could use a little refactoring (the
lambada component is getting big enough to split).