Avoid skip level headings
My blog page and projects page had a top level `h1` then went directly
to an `h3`. This skip level nesting can make it harder to navigate sites
for accessibility tools.

This commit changes this to be a `h2` instead. No considerations for the
larger size have been taken (it looks fine).