Setup syntax highlighting
This commit adds syntax highlighting to code blocks. It uses a
`remark-prism` plugin which does the highlighting.

 ## Issues with this

One issue I am having with it right now is that the are quotes in "unknown"
languages (unknown to Prism). For example one block reads:

```
syntax = &quot;proto3&quot;;

package messages;

message ThermostatState {
    string name = 1;
    double air_temp = 2;
    double rad_temp = 3;
}
```