Change method of generating RSS feed.
Previously we were using a `page` with `getServerSideProps` to generate
the RSS feed for the blog. This has the issue of requiring a lambda and
making the site not really a static site (only because the `res` object
is abilable for `getServerSideProps` and not `getStaticProps`).

This commit uses a wholly different method of generating a file into the
`public` directory in a `prebuild` step. This should run ahead of
building the site due to the "pre and post scripts" feature of npm:
https://docs.npmjs.com/cli/v8/using-npm/scripts#pre--post-scripts

This change required installing `ts-node` and configuring it to work
using a `commonJS` method of loading dependencies.

We ignore the generated file because we don't want to commit something
that changes frequently to version control.