Update all of the colours with the `prefer-color-scheme: dark` media query.
This commit updates the use of all of the colours to be "light" and
"dark". The dark was created by flipping the bright and the dark and
finding an new link colour.

To find the new link colour the previous link colour `#1565c0` was taken
to https://webaim.org/resources/contrastchecker/ and I slide the
"lightness" slider until the Contrast ratio reached AAA on the "normal
text". Using this I also darked my link colour on the "light" styles.

All of the colours are now looking good in dark mode except for the code
blocks.