Remove the RSS feed link from non-blog pages
Previously the footer contained a link to the RSS feed for every page.
This was a little weird on the projects page where it would be
conceivable to have a feed that included updates to items there.

This achieves this by creating a `BlogLayout` which passes a long a new
prop to the footer.