Correct the port that we use for the base url.
When we switched to Next.js the dev port would have switched to 3000 and
whatever was before was on 8000.

I figured it out from this message:

```
[Jest Dev server] ready - started server on 0.0.0.0:3000, url: http://localhost:3000
[Jest Dev server] event - compiled client and server successfully in 3.7s (176 modules)
[Jest Dev server] Attention: Next.js now collects completely anonymous telemetry regarding usage.
This information is used to shape Next.js' roadmap and prioritize features.
You can learn more, including how to opt-out if you'd not like to participate in this anonymous program, by visiting the following URL:
https://nextjs.org/telemetry

Server has taken more than 45000ms to start.

☝️ You can set "server.launchTimeout" in jest-puppeteer.config.js

```