Add an extra `build` step
This is run even though the test script will build the site in order to
try to speed up the test step (since it should use some build cache? Bit
of a guess that this will work).

Remove the `NODE_ENV` because we were getting warnings and now were
using `start` rather than `dev` for the tests.