Add a canonical link to the header
I am seeing some warnings in Google related to not having a user
selected canonical on a page.

This commit adds this by introducing a utility to get the full url of
the page. We then use this in both the site map and the seo component
used in the layout of the page.